backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
python-multipart
aiofiles
